Issue #547: Added a Sliding Time Window implementation to measure metrics.

+++ resilience4j-core/src/main/java/io/github/resilience4j/core/metrics/SlidingTimeWindowMetrics.java
@@ -35,7 +35,7 @@ public class SlidingTimeWindowMetrics implements Metrics {
      *
      * @param timeWindowSizeInSeconds     the window time in seconds
      */
-    public SlidingTimeWindowMetrics(short timeWindowSizeInSeconds) {
+    public SlidingTimeWindowMetrics(int timeWindowSizeInSeconds) {
         this(timeWindowSizeInSeconds, Clock.systemUTC());
     }
